requirements.txt
Flask
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
google-adk
google-generativeai
pymongo
dnspython